first commit on tuidoku
The plan is to create a fork of my program console_sudoku
to add a few features, such as a UI to generate puzzles with
different numbers of given squares, or to play from a puzzle
file. The program will also have features such as generating
puzzles either as a printable format or machine readable. Also
will have the ability to solve given puzzles.