made selection window and files return simpleboard
This is because it was taking a very long time to parse
files with a large amount of puzzles in them since they
were all being solved. This way solving the puzzle takes place
after the puzzle has been selected.

I noticed that since the selection window and the game window
are seperate ncurses windows with endwin() called in between and
it takes a split second to solve so for a split second it is
possible to see the text present before ncurses was started